feat: faucet page redesign
